# product-microservice/src/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"ErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {
						"type": "boolean",
						"enum": [
							"true"
						],
						"nullable": false
					}
				},
				"required": [
					"message",
					"error"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ms_node_1",
		"version": "0.1.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/events/list/{type}/{id}": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"items": {
												"properties": {},
												"type": "object"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "type",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"Product"
							]
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}